refactor: lens-debate — Internal-link resolution split across two package homes for one volatility axis
Moved resolveWikilink + WikilinkResolution from client/ui/lineRef.ts into @kolu/solid-browser's relativePath.ts (reusing posixDir, adding a basename helper), exported it from the package index, repointed BrowseFileDispatcher to the package front door, and relocated the test block.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
