+++ packages/solid-browser/src/index.ts
@@ -8,5 +8,10 @@ export {
   type CreateBrowserOptions,
   DEFAULT_MAX_ENTRIES,
 } from "./createBrowser";
-export { resolveLinkHref, resolveRelativePath } from "./relativePath";
+export {
+  resolveLinkHref,
+  resolveRelativePath,
+  resolveWikilink,
+  type WikilinkResolution,
+} from "./relativePath";
 export { type PreviewPathCodec, pathFromPreviewPathname } from "./previewPath";
